Constants not available with CMake minimal builds

+++ src/rmt.rs
@@ -211,22 +211,22 @@ pub fn duration_to_ticks(ticks_hz: Hertz, duration: &Duration) -> Result<u16, Es
     let ticks = duration
         .as_nanos()
         .checked_mul(u32::from(ticks_hz) as u128)
-        .ok_or_else(|| EspError::from(ERR_EOVERFLOW as i32).unwrap())?
+        .ok_or_else(|| EspError::from(ERR_EOVERFLOW).unwrap())?
         / 1_000_000_000;
 
-    u16::try_from(ticks).map_err(|_| EspError::from(ERR_EOVERFLOW as i32).unwrap())
+    u16::try_from(ticks).map_err(|_| EspError::from(ERR_EOVERFLOW).unwrap())
 }
 
 /// A utility to convert ticks into duration, depending on the clock ticks.
 pub fn ticks_to_duration(ticks_hz: Hertz, ticks: u16) -> Result<Duration, EspError> {
     let duration = 1_000_000_000_u128
         .checked_mul(ticks as u128)
-        .ok_or_else(|| EspError::from(ERR_EOVERFLOW as i32).unwrap())?
+        .ok_or_else(|| EspError::from(ERR_EOVERFLOW).unwrap())?
         / u32::from(ticks_hz) as u128;
 
     u64::try_from(duration)
         .map(Duration::from_nanos)
-        .map_err(|_| EspError::from(ERR_EOVERFLOW as i32).unwrap())
+        .map_err(|_| EspError::from(ERR_EOVERFLOW).unwrap())
 }
 
 pub type TxRmtConfig = config::TransmitConfig;
@@ -781,7 +781,7 @@ impl<const N: usize> FixedLengthSignal<N> {
         let item = self
             .0
             .get_mut(index)
-            .ok_or_else(|| EspError::from(ERR_ERANGE as i32).unwrap())?;
+            .ok_or_else(|| EspError::from(ERR_ERANGE).unwrap())?;
 
         // SAFETY: We're overriding all 32 bits, so it doesn't matter what was here before.
         let inner = unsafe { &mut item.__bindgen_anon_1.__bindgen_anon_1 };
